In-game shop created (accessible from menu)
files edited:
MenuActivity: added intent for new shop button
ShopActivity: added buttons for categories of available items, and scrollview that contains those items. (Items under scrollview will be updated when player clicks on a different category.)

new files:
ShopItems: abstract class for new classes GameShop, PixelShop, TileShop and GpaShop, which contains the images and their descriptions for the available items in store.

ShopItemsBuilder: responsible for setting the new images and descriptions of each item for the chosen category of items. Called by ShopActivity when a player clicks on a button at the top of the screen.